# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChowHub Ghana - Food & Dining Discovery Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/update:
    patch:
      operationId: updateUserProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Vendor Auth ──
  /auth/vendor/register:
    post:
      operationId: registerVendor
      tags: [auth]
      summary: Register a new vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterVendorBody"
      responses:
        "201":
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorAuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/vendor/login:
    post:
      operationId: loginVendor
      tags: [auth]
      summary: Login vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorAuthResponse"
...
  /auth/admin/login:
    post:
      operationId: loginAdmin
      tags: [auth]
      summary: Login admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # ── Listings (Public) ──
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth Schemas ──
    RegisterUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        city:
          type: string
      required: [name, email, phone, password, city]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        city:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    VendorAuthResponse:
      type: object
      properties:
        token:
          type: string
        vendor:
          $ref: "#/components/schemas/Vendor"
      required: [token, vendor]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required: [id, name, email, phone, city, createdAt]
    RegisterVendorBody:
      type: object
      properties:
        businessName:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
      required: [businessName, email, phone, password]
    Vendor:
      type: object
      properties:
        id:
          type: string
        businessName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        plan:
          type: string
        planExpiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, businessName, email, phone, status, plan, createdAt]

    # ── Listing Schemas ──
...