# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChowHub Ghana - Food & Dining Discovery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  name: string;
  email: string;
  phone: string;
  city: string;
  role?: string;
  createdAt: string;
}